Fixes #93

+++ web/src/model/auth.ts
@@ -1,39 +1,47 @@
 import { PayloadAction } from "@reduxjs/toolkit";
 import { AuthProvider, createUserWithEmailAndPassword, deleteUser, EmailAuthProvider, getAuth, sendEmailVerification, signInWithEmailAndPassword, signInWithPopup, signOut, updateProfile, User, UserCredential } from "firebase/auth";
+import { doc, getDoc, getFirestore } from "firebase/firestore";
 import { batch } from "react-redux";
 import { Action } from "redux";
+import EntityApi from "../fbase/EntityApi";
+import EntityClient from "../fbase/EntityClient";
+import { mutate, setAuthUser } from "../fbase/functions";
+import LeaseeApi from "../fbase/LeaseeApi";
 import authEmailVerified from "../store/actions/authEmailVerified";
 import authRegisterStageUpdate from "../store/actions/authRegisterStageUpdate";
 import authSessionBegin from "../store/actions/authSessionBegin";
 import authSessionNameUpdate from "../store/actions/authSessionNameUpdate";
 import { AppDispatch, RootState } from "../store/store";
 import { logError } from "../util/common";
+import { alertError, alertSuccess, setSuccess } from "./alert";
 import { deleteOwnedDecks } from "./deck";
 import firebaseApp from "./firebaseApp";
-import { checkUsernameAvailability, createIdentity, deleteIdentity, getIdentity, replaceAnonymousUsernameAndDisplayName, setAnonymousIdentity, setNewIdentity } from "./identity";
+import { IDENTITIES } from "./firestoreConstants";
+import { checkUsernameAvailability, createIdentity, deleteIdentity, getIdentity, replaceAnonymousUsernameAndDisplayName, setAnonymousIdentity, setNewIdentity, watchCurrentUserIdentity } from "./identity";
+import { appGetState } from "./lerni";
 import { createFirestoreLibrary, deleteLibrary, saveLibrary } from "./library";
-import { ANONYMOUS, GET_IDENTITY_FAILED, Identity, IDENTITY_NOT_FOUND, INFO, LerniApp, RegisterStage, REGISTER_BEGIN, REGISTER_EMAIL, REGISTER_EMAIL_USERNAME_RETRY, REGISTER_EMAIL_VERIFY, REGISTER_PROVIDER_USERNAME, Session, SessionUser, SIGNIN_FAILED, UserNames } from "./types";
+import { ANONYMOUS, GET_IDENTITY_FAILED, Identity, IDENTITY_NOT_FOUND, INFO, LerniApp, LerniApp0, SignInResult, SIGNIN_OK, RegisterStage, REGISTER_BEGIN, REGISTER_EMAIL, REGISTER_EMAIL_USERNAME_RETRY, REGISTER_EMAIL_VERIFY, REGISTER_PROVIDER_USERNAME, Session, SessionUser, SIGNIN_FAILED, UserNames } from "./types";
 
-export function doAuthRegisterBegin(lerni: LerniApp, action: Action<string>) {
+export function doAuthRegisterBegin(lerni: LerniApp0, action: Action<string>) {
     lerni.authRegisterStage = REGISTER_BEGIN;
 }
 
-export function doAuthRegisterCancel(lerni: LerniApp, action: Action<string>) {
+export function doAuthRegisterCancel(lerni: LerniApp0, action: Action<string>) {
     endRegistration(lerni);
 }
 
-function endRegistration(lerni: LerniApp) {
+function endRegistration(lerni: LerniApp0) {
     delete lerni.authRegisterStage;
 }
 
-export function doAuthEmailVerified(lerni: LerniApp, action: PayloadAction) {
+export function doAuthEmailVerified(lerni: LerniApp0, action: PayloadAction) {
     const user = lerni.session?.user;
     if (user) {
         delete user.requiresEmailVerification;
     }
 }
 
-export function doAccountDeleteEnd(lerni: LerniApp, action: PayloadAction<boolean>) {
+export function doAccountDeleteEnd(lerni: LerniApp0, action: PayloadAction<boolean>) {
     delete lerni.session;
     lerni.alertData = {
         message: "Your account has been deleted",
@@ -41,29 +49,16 @@ export function doAccountDeleteEnd(lerni: LerniApp, action: PayloadAction<boolea
     }
 }
 
-export function doAuthSignin(lerni: LerniApp, action: PayloadAction<boolean>) {
-
-    if (action.payload) {
-        lerni.signinActive = true;
-    } else {
-        delete lerni.signinActive;
-    }
-}
-
-export function doAuthSignout(lerni: LerniApp, action: PayloadAction) {
-    delete lerni.session;
-}
-
-export function doAuthSessionEnd(lerni: LerniApp, action: Action) {
+export function doAuthSessionEnd(lerni: LerniApp0, action: Action) {
     delete lerni.session;
 }
 
 
-export function doAuthRegisterStageUpdate(lerni: LerniApp, action: PayloadAction<RegisterStage>) {
+export function doAuthRegisterStageUpdate(lerni: LerniApp0, action: PayloadAction<RegisterStage>) {
     lerni.authRegisterStage = action.payload;
 }
 
-export function doAuthSessionNameUpdate(lerni: LerniApp, action: PayloadAction<UserNames>) {
+export function doAuthSessionNameUpdate(lerni: LerniApp0, action: PayloadAction<UserNames>) {
 
     const session = lerni.session;
     if (session) {
@@ -155,7 +150,7 @@ export async function deleteUserData(userUid: string) {
 
 
 
-export function doAuthRegisterEmailVerified(lerni: LerniApp, action: PayloadAction) {
+export function doAuthRegisterEmailVerified(lerni: LerniApp0, action: PayloadAction) {
     endRegistration(lerni);
     lerni.alertData = {
         severity: INFO,
@@ -163,22 +158,12 @@ export function doAuthRegisterEmailVerified(lerni: LerniApp, action: PayloadActi
     }
 }
 
-
-export function selectSession(state: RootState) {
-    return state.lerni.session;
-}
-
-/** Select the currently signed in user or undefined if no user is signed in */
-export function selectCurrentUser(state: RootState) {
-    return state.lerni?.session?.user;
-}
-
 export function selectRegistrationState(state: RootState) {
     return state.lerni.authRegisterStage;
 }
 
-export function selectSigninActive(state: RootState) {
-    return Boolean(state.lerni.signinActive);
+export function selectSigninActive(lerni: LerniApp) {
+    return Boolean(lerni.signinActive);
 }
 
 export function selectAccountIsIncomplete(state: RootState) {
@@ -216,6 +201,23 @@ export function createEmptySession() {
     return result;
 }
 
+export function createSessionUser(user: User, identity: Identity) {
+    const requiresEmailVerification = getRequiresVerification(user);
+
+    const result: SessionUser = {
+        uid: user.uid,
+        username: identity.username,
+        displayName: identity.displayName,
+        providers: getProviders(user)
+    }
+
+    if (requiresEmailVerification) {
+        result.requiresEmailVerification = requiresEmailVerification;
+    }
+
+    return result;
+}
+
 export function createSession(
     uid: string, 
     providers: string[],
@@ -236,7 +238,9 @@ export function createSession(
     return {user}
 }
 
-export function doAuthSessionBegin(lerni: LerniApp, action: PayloadAction<Session>) {
+
+
+export function doAuthSessionBegin(lerni: LerniApp0, action: PayloadAction<Session>) {
     lerni.session = action.payload;
 }
 
@@ -244,8 +248,38 @@ export function getProviders(user: User) {
     return user.providerData.map(data => data.providerId);
 }
 
+export async function providerRegister(client: EntityClient, provider: AuthProvider) {
+    const auth = getAuth(firebaseApp);
 
-export async function providerRegister(dispatch: AppDispatch, provider: AuthProvider) {
+    // `signInWithPopup` throws an exception if sign-in fails.
+    const result = await signInWithPopup(auth, provider);
+    const user = result.user;
+    const uid = user.uid;
+    const displayName = user.displayName || ANONYMOUS;
+    // const username = ANONYMOUS;
+
+    // const identity: Identity = {
+    //     uid,
+    //     username,
+    //     displayName
+    // }
+
+    // const sessionUser = createSessionUser(user, identity)
+   
+    const lib = createFirestoreLibrary();
+    await Promise.all([
+        saveLibrary(uid, lib),
+        setAnonymousIdentity(uid, displayName)
+    ])
+
+    mutate(client, (lerni: LerniApp) => {
+        lerni.authRegisterStage = REGISTER_PROVIDER_USERNAME;
+    })
+
+}
+
+
+export async function providerRegister0(dispatch: AppDispatch, provider: AuthProvider) {
     
     const auth = getAuth(firebaseApp);
 
@@ -272,7 +306,35 @@ export async function providerRegister(dispatch: AppDispatch, provider: AuthProv
     })
 }
 
-export async function emailPasswordSignIn(email: string, password: string) {
+export async function emailPasswordSignIn(api: EntityApi, email: string, password: string) {
+    
+    const auth = getAuth(api.getClient().firebaseApp);
+    let credential: UserCredential | null = null;
+    try {
+        credential = await signInWithEmailAndPassword(auth, email, password);
+    } catch (error) {
+        console.error(error);
+        return SIGNIN_FAILED;
+    }
+    const user = credential.user;
+
+    let identity: Identity | null = null;
+    try {
+        identity = await getIdentity(user.uid);
+    } catch (error) {
+        console.error(error);
+        return GET_IDENTITY_FAILED;
+    }
+    if (!identity) {
+        return IDENTITY_NOT_FOUND;
+    }
+    const sessionUser = createSessionUser(user, identity);
+    setAuthUser(api.getClient(), sessionUser);
+    alertSuccess(api, "Welcome back!");
+    return SIGNIN_OK;
+}
+
+export async function emailPasswordSignIn0(email: string, password: string) {
     
     const auth = getAuth(firebaseApp);
     let credential: UserCredential | null = null;
@@ -289,7 +351,6 @@ export async function emailPasswordSignIn(email: string, password: string) {
     } catch (error) {
         throw new Error(GET_IDENTITY_FAILED, {cause: error});
     }
-    debugger
     if (!identity) {
         await deleteUser(user).catch(
             error => console.error(error)
@@ -301,41 +362,51 @@ export async function emailPasswordSignIn(email: string, password: string) {
     return createSession(user.uid, providers, identity.username, identity.displayName, requiresVerification);
 }
 
-export async function providerSignIn(provider: AuthProvider) {
-    
-    const auth = getAuth(firebaseApp);
+export function endSignIn(client: EntityClient) {
+    const lerni = appGetState(client);
+    if (lerni.signinActive) {
+        mutate(client, (lerni: LerniApp) => {
+            delete lerni.signinActive;
+            setSuccess(lerni, "Welcome back!");
+        })
+    }
+}
 
+
+export async function providerSignIn(api: EntityApi, provider: AuthProvider) : Promise<SignInResult> {
+    const firebaseApp = api.getClient().firebaseApp;
+    const auth = getAuth(firebaseApp);
     let result: UserCredential | null = null;
     try {
         result = await signInWithPopup(auth, provider);
     } catch (error) {
-        throw new Error(SIGNIN_FAILED, {cause: error});
+        console.error(error);
+        return SIGNIN_FAILED;
     }
     const user = result.user;
-    const uid = user.uid;
-    const displayName = user.displayName || ANONYMOUS;
-    const providers = getProviders(user);
-    const requiresVerification = getRequiresVerification(user);
-
-    let identity: Identity | null = null;
+    const db = getFirestore(firebaseApp);
+    const idRef = doc(db, IDENTITIES, user.uid);
     try {
-        identity = await getIdentity(uid);
+        const idDoc = await getDoc(idRef);
+        if (!idDoc.exists()) {
+            await deleteUser(user).catch(
+                error => console.error(error)
+            )
+            return IDENTITY_NOT_FOUND;
+        }
+        const identity = idDoc.data() as Identity;
+        const sessionUser = createSessionUser(user, identity);
+        setAuthUser(api.getClient(), sessionUser);
+        alertSuccess(api, "Welcome back!");
+        return SIGNIN_OK;
     } catch (error) {
-        await deleteUser(user).catch(
-            error => console.error(error)
-        );
-        throw new Error(GET_IDENTITY_FAILED, {cause: error});
+        console.error(error);
+        return GET_IDENTITY_FAILED
     }
 
-    if (identity === null) {
-        await deleteUser(user).catch(
-            error => console.error(error)
-        );
-        throw new Error(IDENTITY_NOT_FOUND);
-    }
 
-    return createSession(uid, providers, identity.username, displayName, requiresVerification);
 }
+
 
 export async function handleAuthStateChanged(user: User) {
 
@@ -391,4 +462,35 @@ export function stopEmailVerificationListener() {
         clearInterval(verificationIntervalToken);
         verificationIntervalToken = null;
     }
+}
+
+export function userTransform(api: LeaseeApi, user: User) {
+    const [, identity] = watchCurrentUserIdentity(api, user.uid);
+    if (identity) {
+        const sessionUser = createSessionUser(user, identity);
+        return sessionUser;
+    }
+    return undefined;
+}
+
+export async function authSignOut(api: EntityApi) {
+    const auth = getAuth(api.getClient().firebaseApp);
+
+    try {
+        await signOut(auth);
+    } catch (error) {
+        alertError(api, "An error occurred during sign out", error);
+    }
+}
+
+export function authBeginSignIn(api: EntityApi) {
+    api.mutate((lerni: LerniApp) => {
+        lerni.signinActive = true;
+    })
+}
+
+export function authEndSignIn(api: EntityApi) {
+    api.mutate((lerni: LerniApp) => {
+        delete lerni.signinActive;
+    })
 }
